move tables into db

+++ src/stats.h
@@ -45,7 +45,7 @@ struct _Level {
 
 // Holds skill information
 struct _Skill {
-	float Data[SKILL_COUNT][2];
+	int Data[SKILL_COUNT][2];
 };
 
 // A single entry for an item drop
@@ -121,7 +121,7 @@ class _Stats {
 		int GetMaxLevel() const { return (int)Levels.size(); }
 
 		int GetValidSkillLevel(int Level);
-		float GetSkill(int Level, int Type, int Index=0) const { return Skills[Level].Data[Type][Index]; }
+		int GetSkill(int Level, int Type, int Index=0) const { return Skills[Level].Data[Type][Index]; }
 		float GetSkillBonusMultiplier(int Level, int Type, int Index=0) const { return (100 + Skills[Level].Data[Type][Index]) * 0.01f; }
 
 		void GetRandomDrop(const _ItemDrop *ItemDrop, _ObjectSpawn *ObjectSpawn);
